requirements.txt
Django
python-dotenv
termcolor
psycopg2
psycopg2-binary
pylint-django
Pillow
django-tempus-dominus
django-widget-tweaks